+++ CHANGELOG.md
@@ -3,6 +3,8 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [0.9.2] - (2019-05-09)
+
 ### Changed
 
 - Revert back to the standard library hashmap because it will use hashbrown very
@@ -119,7 +121,8 @@ vector of results (thanks @ovarene)
 ### Added
 - Implement Hamming, Jaro, Jaro-Winkler, and Levenshtein
 
-[Unreleased]: https://github.com/dguo/strsim-rs/compare/0.9.1...HEAD
+[Unreleased]: https://github.com/dguo/strsim-rs/compare/0.9.2...HEAD
+[0.9.2]: https://github.com/dguo/strsim-rs/compare/0.9.1...0.9.2
 [0.9.1]: https://github.com/dguo/strsim-rs/compare/0.9.0...0.9.1
 [0.9.0]: https://github.com/dguo/strsim-rs/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/dguo/strsim-rs/compare/0.7.0...0.8.0